Clean up array subscript has type ‘char’ warnings.

+++ src/libCom/yacc/error.c
@@ -52,7 +52,7 @@ print_pos(char *st_line, char *st_cptr)
     if (st_line == 0) return(0);
     for (s = st_line; *s != '\n'; ++s)
     {
-	if (isprint(*s) || *s == '\t')
+	if (isprint((int) *s) || *s == '\t')
 	    putc(*s, stderr);
 	else
 	    putc('?', stderr);
